Agaent bootstrap type
Following on from #16313 this introduces a type for agent bootstrap.
It removes the need to compose the init state args and just flattens
it to what you actually require. There are args which are required
and optional args. Not everything needs to set optional arguments.
We could merge args and optional values as well to make it cleaner.

Also the initialize function is still huge, but we can tackle that
another day. For now this should make it simpler to organise things.