Add validate and IsLocal to CharmOrigin.
Validating a CharmOrigin for state is complex. Add a method that knows
the details and document them. Add IsLocal method as well to ensure we
choose the correct Match method. Especially with "ch" vs "charm-hub",
charm schema vs charm source, they are not compatible. "local" == "local"
makes it easier to get the wrong Matching for other values.