Refactor service factory model getter
We already have a way to get a watchable db from a namespace. The
namespace here can either be the "controller" namespace or a model
uuid. Removal of the watchable db for model factory isn't required
if we already know the model uuid.

I've also gone and removed the ModelUUID method from the model
factory. We need a method on the model factory to ensure that the
ModelFactory interface doesn't satisfy against every type. Instead
of ModelUUID I've added a TODO() method. This will be removed once
we add methods to the interface.

The ServiceFactoryGetterFunc was also lifted out of the service
factory interfaces package and into the domain/servicefactory/testing
suite. We don't want people to depend on this and hiding it away in
the testing suite should cover the use case it was designed for.

----

It should be noted that if we do need ModelUUID then there is a
problem with the modeling. I suspect most of the ModelUUID code is
either at the apiserver root, which is required for getting the
model first time around or in state. If it's the latter, that will
not be required as you're already acting upon the existing model.

Anything outside of that is incorrect and will need to be solved
in some other way.