Drop json codec v0 support
The following drops v0 json codec support. This is ongoing work to
support tracing between client and server requests. I don't believe
we should be supporting v0 anymore with the next major release (v4),
so before I add a TraceID I thought it prudent to clean up the code.

We would need to ensure that pylibjuju also sends a Version via the
header (Version=1), so we don't break compatibility.

Also for v0 messages we were double unmarshalling to detect the
version, via the RequestId. If it was 0, it was assumed that it
was incorrect and would attempt to unmarshall again. Although, good
for backwards compatibility, not great for efficiency.