Export revision -1 if revision is nil
Previously when exporting a charm origin without a revision, we would
default it's value to 0.

However, 0 is a valid charm revision. This means we can no longer tell
the difference between a charm origin without a revision, and an origin
with revision 0

Instead, default to revision -1, which across our codebase is used as
another way to denote no revision