RPC supports trace id and span id
The following implements the w3c tracing context document[1] for
both trace-id[2] and span-id[3]. Sometimes span-id is called
parent-id, but for out setup we're going to be explicit and call it
span-id like the document allows.

Currently, there is no validation of the values of a trace-id or
a span-id. At the moment, I'm thinking we can ignore that and force
the OTEL (open telemetry) clients force that validation. Most likely,
they'll be the ones seeding that information.

As tracing is optional, I don't want to bring down Juju for a bad
trace and span id.

 1. https://www.w3.org/TR/2021/REC-trace-context-1-20211123/
 2. https://www.w3.org/TR/2021/REC-trace-context-1-20211123/#trace-id
 3. https://www.w3.org/TR/2021/REC-trace-context-1-20211123/#parent-id