Use existing controller config from args
To deploy the controller charm we need controller config, which we
already have, so there is no need to call it again from state. This
is a waste. I believe we could easily do this for all the other things,
we do during bootstrap. We just need to use the minimum information
and not the maximum.