Removes ControllerConfig calls from state tests
To help find the wood from the trees, the number of calls to
ControllerConfig in tests are quite large. By removing the calls
sites, we can actually see what is calling ControllerConfig. The
next set of commits/PRs will hopefully just reduce state calls
to the PeerGrouper and APIServer workers. We can't remove those
because of the chicking and egg problem. In that we don't have a
db yet (at least not in HA) and attempting to get the controller
config isn't yet possible.

I removed a few tests:

 1. Watching controller config test is replaced via the domain
    package.
 2. Tests to get and update controller config are covered by the
    domain package.

All the tests can rely on the FakeControllerConfig with the addition
of the management space.