Agent controller config subscription
Allow a subscription model based on a dispatched signal if a SIGHUP
is sent. The following leans heavily on the worker runner type. It
allows the lifecycle management of a subscription. If the subscription
is unsubscribed then we just kill the tomb, the runner will manage
the clean up.

It is expected that every downstream consumer should read off the
changes channel as quickly as possible. Any hold up will cause
back pressure on the rest of the consumers, as new events will
not be processed.

This is similar to the change stream, but we can be a lot less complex
as we don't need to filter on on patterns. If that does come a
requirement, which I don't suspect will happen, then we should crib
off that one.

In order to provide this functionality, then a new change to the
runner type is required[1]. This change allows us to see what
workers are within the runner (see PR for more details).

Note: it is expected that the consumer should check the Receive
operator[2] to see if the config worker has been closed. It is best
practice if that happens, to bounce the worker as the agent controller
config is dying. This can be ignored if the unsubscribed has been
called and return gracefully.

Until [1] has been landed, I've hardcoded the go.mod.

 1. https://github.com/juju/worker/pull/32
 2. https://go.dev/ref/spec#Receive_operator