Remove docker checks on controller config
For some reason we validated against the docker image repo everytime
we validated the contents. The problem is that if you had a value
set, then every validation will request to docker. This was a waste
of resources.

It also breaks expectations. No getter should have a side effect. This
was even worse if we're rate limited, or there is a denial of service,
preventing us to use the controller.

I'm now having to unwind this because of bad judgement.