Ensure charm origins have a revision
This will be useful for a future change, since we will now be able to
use the charm origin to get a revision, rather than being force to parse
it from the charm url

The client always fills in a revision when deploying. Also the apiserver
parses from the provided charm url is a revision is not provided. This
is for compatibility with old clients. In future, we will likely enforce
an origin is provided in validation steps

Add upgrade steps which add a revision to charm origins if it doesn't
have one. Parsed from the charm url

Do the same for migration on import

As a flyby, on migrations make sure to export -1 as the revision for charm
origins without a revision, because 0 is a valid revision for local
charms.