Allow DB accessor to be more resilient
The following is a thought experiement from the london sprint, which
suggested we could be more resilient to failure to open the db. As
we previously only opened the db, that might not actually be enough.
We might require to re-open the db handler and try again. This
attempts to do this without introducing the dbApp directly into the
tracker.