Introspection: reload config via SIGHUP
The following allows sending a SIGHUP to the jujud process. This
will be useful for testing and may also be used by the juju-controller
charm (although I suspect we'll do that in python directly).

This will also target non-controller jujud processes, but that will
be treated as a no-op. If we move to controller-jujud process, then
we can grab that pid, which will be safer.