Ensures peergrouper does not eschew adding controller trackers for machines with "pending" status.
There is no need to do this, as:

1. The replicaset is only updated when new servers can be contacted.

2. API server details will only ever have addresses for machines after
they have been set for the first time anyway.

This mitigates a race where we tell a controller to use localhost for
the API address, but there is no API available there because we
restarted the agent as a controller before it set its addresses for the
first time.