Remove stateservinginfo from peergrouper
This was observed when identifying what we require for the new
controller agent config. As part of undertaking that work we need
to remove APIPort and ControllerAPIPort (state port will follow),
as it doesn't need to exist in the new world.

The peergrouper has state, so just use the controller config. If
it has the controller config, it will have all the relevant
information.

I've removed the check for missing state serving info, as that isn't
the case anymore.