Machiner state converter
The state converter worker and the machiner get tied into knots,
as there is a edge case about ensuring the state. The following
moves the state converter worker into the machiner. Removing the
state converter worker will be done in a subsequent commit.

The machiner is the best place to do this and it just replicates
what we have currently.