Remove the IsNotControllerFlag
This is removed as nobody is using it now we've removed the
state converter worker. I'm still unsure this will do exactly
as we want and the reason to have the two workers was to prevent
the bootstrap node querying when it doesn't need to.

If this does what we want, it removes some complicated flag logic.