Thread context.Context into APICaller
The following is a lot of mechanical changes to thread
context.Context into the APICaller. This will allow passing in the
tracing span ids into this, so we can then pass it on to the facades.

It also allows for better cancelation properties.

The next phase after this one, is to allow the FacadeCall to take
a context.Context. That's more work, but will fully expose a
context.Context to the API.