Fix notify sighup tests
As the sighup tests fail because the tests grabed the process of
the current test, but that might not be the right thread. We
could use runtime lock to thread, but that doesn't seem ideal. Instead
we move the notify logic out of the worker and allow a test one
to inject.