Fix bug where "use --force" prompt is not displayed
For some reason our series selector was not returning an
UnsupportedSeries error when it should have been, instead duplicating
the text of the error and returning a generic error

Return an UnsupportedSeries error when the requested series is not
supported by the charm. This can then be detected downstream.

However, if the series is not supported by Juju, then we return a normal
NotSupported error

Remove some duplicated checks. We do not need validate the output of
charmSeries, since the series selector guarentees the returned series is
supported by juju

Drop selecting from charm url as this is a charmstore only feature

We also drop charmValidationError in a few places, because all this did
was reformat the error from charmSeries. Now I have re-written these
error messages, this is not longer required