Better fetch instance support.
Juju's ec2 provider has always done a fetch of instance type names with
an AZ filter and then taken those results and queried the region based
on the results.

We have witnessed errors in this approach where the results from one AWS
call are in conflict with DescribeInstanceTypes.

Because our AWS client is already scoped to the region of the model
cloud we don't need to do this double handling. Instead we can just
query the region directly for instance types.

This removes the potential for conflicting data scenarios.