Add system identity public key to authorized_keys on new model configs
This patch aims to fix https://bugs.launchpad.net/juju/+bug/1834974
where we cannot apply iptables rules on OCI.
This happens because on OCI (and Vsphere and Equinix) we use the
SshInstanceConfigurator which applies iptables' rules via ssh, but on
the units the authorized_keys file under /home/ubuntu/.ssh is not fully
populated: the system identity public key is missing and therefore the
controller machine cannot ssh onto it.

We fix this by adding this missing key to the model config at model
creation.