Refactor the upgrades service to use statemachine
The following re-writes the upgrades service to use a statemachine
so it makes it easier to know that you can't jump from created
to done without doing the intermidate steps.

The transactions are modeled directly in the db. This makes it
easier to understand what state the upgrade is at without the implicit
knowledge of completed_at time present or not. By making it explicit
we can observe those changes easily via a watcher.

In addition, we enforce a upgrade UUID on the service. This is to
prevent the string, string hell. Types provide a way to differentiate
what we expect to be sent.