Predicate value watcher
The predicate value watcher allows you to prevent changes from being
notified out if the predicate returns false.

This should prevent the custom watchers that we're collecting in the
state package. We want to keep everything lean as possible and prevent
custom watchers from being implemented and they're all different.