Wire up dqlite trace
The following injects the trace into the dqlite library via the
dqlite tracing package.

We only inject the trace if a tracer is enabled. If it's not enabled
we just skip the unwanted work. In addition the dqlite tracing has
a smaller surface area of a tracing library. The dqlite library
doesn't need to implement the whole of the OTEL libary as it's
not currently tracing themselves. This might change in the future,
but for now it's limited (in a good way).
With this in mind we pool as much of the shimmed types as we're going
to be making thousands of these types.