Use a simple select query for lease expiry
The following reduces the number of writes we perform for expirying
leases. First select if there are any leases that meet the criteria
and if that does match, then perform the delete. The read transaction
will be promoted to a write when required. This should prevent any
potential write contention.