Fix some of the issues around start upgrade
If we're already at the start upgrade step, then no amount of
calling it will start the upgrade.

We have to remember that if a step has failed, it should not be
possible to undo that, without manual intervention; we don't know
what the state of the world is at. If we can't progress this action,
then there is no automatic way to undo this.