Scaffolding for the objectstore
The following scaffolds out the objectstore. The objectstore will
work as a dependency to the api server. The api server will then
request a new obejct store for a given namespace. We can back that
by mongo at first, but we can then use in-memory/file/s3 backing
when required.