Fix aws_ami_creation instance_type
It turns out the instance_type t2.medium can only be used for x86_64 and
i386 architectures. Perhaps this was a recent change? Has this ever
worked for arm64?

Add a select-case clause to select an appropriate instance_type base on
the architecture. We only need amd64 and arm64 for the time being