Implement predicate watcher for namespaces
The namespace watcher should also have an associated predicate. We
should do this at the same time to prevent any incorrect
implementations in the future.

The incorrectness would come from the fact that you might want to
ignore the initial change, but that's incorrect. You should always
send it, as that's a requirement for the API.