Close temp files before potential errors
The following ensures that we close files before an error is returned.
This was spotted just whilst auditing some code for future estimates.

In most cases the io.Copy could return an error and in those cases,
we should close the tmp file before returning. This should ensure
cleanup of resources.