Wire up the state object store
This object store is purely based on the state object store. It
provides backward compatibility with the existing storage provider
i.e. mongo.

New storage providers will be added in the future, but to get things
off the ground lets remove the api dependency on state storage
directly.

Once we've proved get works, we can then move on to put etc.