Use fsnotify instead of inotify
It was brought up that we shouldn't depend on k8s utilities for
non k8s infrastructure if there was another solution. The other
solution was fsnotify, which is almost identical to inotify. The
only difference from the perspective that we want is that it has
a different mask type.

The rest is nicely abstracted away because we don't leak our types,
it's all self contained in the worker.