Watch upgrades if it's already started
If the upgrade has already started we want to watch the active
upgrade. We can utlise the new predicate watcher here, by just
watching for a state to change.

There is a slight complication. If we can't locate an active upgrade
then we're left in a limbo land. We either just missed it (not sure
how, as we expect everyone to participate) or something went wrong.
The only course of action is to either force the whole re-evaluation
or die. I chose the former.

I also changed the return error types for the worker. As this is
the upgrade worker, it should only ever evaluate once we start up.
There is never a point that we want to reconsider this worker after
that. In that case, we can just uninstall ourselves and call it done.
This should reduce the number of workers to consider as dependencies
from the dependency graph and potential goroutines.