Clear hash, id and channel when switching to local charms
Local charm origins can not include hashes, ids or channels. However,
when switching from a charmhub charm to a local one these attributes
were accidentally copies over

Ensure these attributes are droped when constructing the new charm
origin

Similarly with the revision, a new revision would not be set, meaning
the revision of the original charm would be kept. Make sure the revision
is updated as well

There is a slight nuance here, simply deploying a local charm does not
fill in a revision, but switching to a local charm will fill in a
revision. This is not a problem, as ideally we would have a revision for
local charms. This is something added in a later version