Expose dqlite node id and model list
The following changes exposes the dqlite node id and model list for
upgrade databases.

The schema wants the dqlite node id to be able to apply the schema
upgrades. Although in theory we can just pass 0, I'm unsure
that's a sound model. Instead we can just expose this, we can then
verify the controller node exists for the dqlite node and don't get
further down the line then what we want.

We also want to be able to list over all the models to apply each
schema update. Exposing the model list will enable us to do that.