Wire up the tracer to the api caller
This is the lightest touch to add a tracer to the api caller. Ideally
you're get it from the context, but adding a context to all the api
calls is a bit daunting. For now we allow passing a tracer to the
client. We can always peel of from the context when we feel brave
enough to change it.