Fix parsing of watch flag on status command
On the status command, when passing the --watch flag, we pass the entire
command line arguments (including 'juju status') along with all the
flags except --watch to viddy, which will run the command indefinitely
with an interval.

To do this, we must remove the --watch flag along with its value from
the list of OS arguments to be passed to viddy. This was incorrectly
done because it only accepted unix-style flags.

The fix is to remove both unix and gnu style when passing the watch
flag.