Role back node ID with in upgrade database worker
We shouldn't require the node id for performing schema updates. To
solve this, we'll remove the inserting of the dqlite node id from
within the schema application. Now we apply it only in bootstrapping
the initial dqlite node. As we need a bootstrapped dqlite node for
tests, we also need to apply it to the controller suite.

This feels much more simpler and in keeping with how we wanted
the upgrading of the database to function.