Pass trace id and span id (scope) through to tracer worker
The following wires up end to end a trace from a client to the server
over rpc. We can't use http as each request is it's own payload. To
get around this, we use a scope. A scope is just a trace id and span
id. Each scope is sent via the request and response. At the rpc
server level we push the scope onto the context and pick it up at
the worker trace level. If there is a scope, we construct a new
span from that information which then forms the parent span.

It's now possible for clients to inject there own tracer into a client
and push that through the stack. You can then visually see the trace
going through the system.