Watch before setting controller ready
If we set the controller ready before we start watching, then our
events _can_ get coalesced into the initial event. So sometimes it
works and sometimes it doesn't. To fix it, only set controller ready
after we've watched. That way we don't run into these issues where
we don't see the event.