Allow integration tests to run against gridfs
The following tests that fail without this patch are actually
full integration tests and are not unit tests. In the future it might
be wise to split these out to correctly handle the complexities of
such.

In the mean time, we just allow serving of files via the gridfs
backing store.