Wire up the apiserver to the charmhub client
The following is all the grunt work of wiring up the charmhub client
to the apiserver. This is a long process as it's used in a lot of places.
Back when I designed the charmhub client I left context.TODO for passing
the tracing library through to it. That time has come to get rid of them.

The code is __very__ mechanical.