Refactor upgrade steps worker
The following refactors the upgrade steps worker to make it easier
to start adding changes to it for the upgrade story. The upgrade
steps worker is an older style worker and we've got new and better
conventions.

The changes are mechanical and are just moving things to better
locations. There is still a horrible wart about using logs as a way
to determine the right path in code. This forces the logs to be
part of the api as it's that what we're testing and not just the
outcome.