Introduce a agent and controller worker
Both the agent and controller versions of the works do similar things,
but not quite the same things. To help improve readability and testing
we split them out into two different workers.

It is much simpler to reason about the logic. Also we're not exposing
dependencies to each worker that don't actually require it. There is
still more work to do to make it more testable.

In addition we can nuke the open state for upgrade worker. That is
just not required. There is a lot more work to do here, like tests
for the controller/agent workers. We need to remove the upgrade state
code and we need to test that it works.

The only wart is the at the preupgrade tests need to expose the
CAAS IAAS enum from state as we need to know upfront if the k8s
upgrade steps can run certain steps.

This does remove a boolean from compile time and puts it upfront
when constructing the dependency engine, akin to a factory pattern.
The wart here is that we're using the state value, which should be in
core.