Rename agent worker to machine worker
The following starts adding more tests to ensure the correct
behaviour is employed. The base worker now has correct steps
to run based on if it's a controller vs a machine.

Most of the mocks are added to because of the amount of interaction
required between the workers and the surounding state.

Update steps themselves will be tested outside of this package.