API client context aware
The following is a very mechanical set of changes, which push
the context type into the facade call methods. This is one more change
to get context everywhere. Unfortunately, it will be a lot of work
to push the context into every method, so for now everything is
just a context.TODO. This will make it a lot easier to fix those
when we do make every method context aware.