Encode controller config for json
The following is a controller config fix whereby any time.Duration
would not be correctly parsed over the wire. The problem occurs
when time.Duration is sent over the wire and we hit the golang
bug. The time.Duration is sent over the wire, but can not be parsed
back out as the time.Duration is a float64. The controller config
doesn't allow coercing float64s instead we opt to encode the
time duration to a string.

This seems a bit like we're papering over the cracks even more. I
feel like the best approach here would be to just allow float64
parsing in the controller config and be done with it.