Insert controller ID on bootstrap
The modelling of the bootstrap via the db concern is wrong. We
MUST always create a db for a controllerNS and the controller model.
It should not be optional to create these databases. The fact that
the concerns are variadic means that isn't enforced at compile time.

This needs to be re-looked at. If this is to fix the dummy cloud
provider type, then I'd rather burn those test to the ground. The
dummy provider needing these hacks just shows that the tests are
breaking assumptions about how things should be laid out.

The variadic concerns need to be explicit for the controllerNS and
controller model. What happens after then can be variadic if we want
to pass additional information for tests (although I'm not happy
with that either).

The fact that a node ID is required to be able to bootstrap shouldn't
be removed.