Remove upgrade logic from state
This is the final piece for upgrades. We no longer provide logic
for upgrades in the state package. Upgade logic should now only be
found in the domain upgrade package.

In doing the following, we have to black hole the AbortModelUpgrade
method. The AbortModelUpgrade is now not supported. There is no way
to rollback in the current implementation. We don't currently have
a way to backup the databases easily and then rollback if it fails.
Nor do we have rollbacks implemented correctly in migrations, so more
thought would be needed.