Remove generic base worker
The concept was to make a generic base worker that provided a tomb
that you could get the kill and wait from. The problem becomes
about how you can also provide the loop/run method. The problem is
that passing the loop is racy and can cause problems. The fix is
to no be generic here, we can solve that later if required.