Thread db lock through to the manifolds
The DB lock was never correctly setup, so we would see issues where
the db would unlock itself, yet the steps worker wouldn't. This is
just painful to debug. The following just forces the lock to be the
same for upgrade steps and databases alike. They just happen to unlock
for different reasons.