Reinstate ifDatabaseUpgradeComplete
When we removed the database worker previously, we removed the
upgrade database flag. We need this to prevent the upgrade steps
from running whilst the upgrade database is running. Essentially,
we need to sequence the upgrade steps against the apiserver and the
upgrade steps require the apiserver to come up. The upgrade database
steps are then a hard dependency on the multiwatcher, which is a
dependency on the apiserver, closing the circle.

It's a bit annoying that it's hard to graph this easily, it would
be nice if there was a way to explain this.