Bug when setting cloud credential with no region.
We introduced a bug into creating models by where if no cloud region was
set on a model we were unable to set a corresponding cloud credential.
This was because we were using the cloud region to validate the cloud a
model was using.

Includes regression test to assert the bug and the fix.