Dqlite safemode
The following implements a non-HA safemode for viewing and editing
the dqlite database without all the ancillary workers. This is
purely for development at the moment. To move into HA safemode, it
will require the other nodes to be co-ordinated and we're currently
blocked on the peer-grouper.

The steps to get the certs is interesting, but once you know the
values then it's possible to do everything over tls.