Cleanup with objectstore
The following just moves the clean up of charms using the object
store. This just ensures that what we create we also use to tear
down.

The next step after this is to push all blobs into the objectstore
and then we can consider backing it with a different storage type
(files and s3).

Note: we're begining to move the object store in to the internal
package, as that's the location of all the different types of backends.
This is because we'll reference the backends in bootstrap and in
the worker packages. This prevents cyclic import issue.