Controller config string duration
The following forces the duration to be a string after coercion. This
is required because once a time.Duration has been serialized via JSON,
the time.Duration is coverted to a float64. There is no way to know if
that time.Duration was the intended goal from the float64 or was actually
intended to be a float64 all along. By using a string we have the ability
to parse that information and validate based purely on the input. It
will either parse or not. This gives a better type to ship over the
wire and we don't loose precision either.