Use network SpaceInfo instead of interfaces where needed
This patch is an scaffolding preparation for the new dqlite domain for
spaces, subnets and AZ.

The idea is to replace the usage of interfaces in favor of the core
network.SpaceInfo in environs and apiserver facades. We need this
because the new service and state in the domain will return core
structs.