Push objecstore through to ToolStorage
The following pushes the objectstorage from the worker to the
tool agent storage requester. The tools is a bit overlay complicated,
as we record the blob metadata in multiple locations. In the
toolsmetadata, managedstoredresources, blobstore collections. Most of
the information is duplicated, it might be wise when we convert the
tools state to the dqlite domain services to reduce the amount of
duplication, or at the very least use RI to remove duplication.

The code changes here, just push the objectstore to the binary storage
type so that it can use the underlying managed storage. I've removed
the layered storage, as we no longer want to have that option of
hierarchical storage for agent tools. Instead just a global namespace
will suffice. In the new world, all the information will be in
the global controller namespace. For the state packages, we use the
controller model-uuid. This will simplify a lot of code we have.

If we can chip away at this complexity as we go, I think it will have
a nice model of what we actually do, rather than an over engineered
view of what we might have wanted at some point.