Fuzz wait-for query and lexer
This is a reactionary PR after a lighting talk at the engineering
sprint. As the wait-for command talks a user input it is advisable
to fuzz the input to prevent a panic from occuring based on the
input.

Fuzzing can take quite sometime we may only want to enable this in
some gated infrastructure (unit gating tests).

This currently hasn't found any issues, but it may well do on
repeated runs.