feat: introduces a new error handling package
This commit starts off a new stem of work for Juju by giving Juju a new
internal errors package so that it can have it's reliance on the current
juju/errors library removed.

This commit brings in a new package called internal/errors for all the
new types. The first change is we have introduced proxying functions for
all the go stderrors helpers. All of these proxied functions are a 1 to
1 mapping the exception that where an error return value is used we have
changed the type to internal/errors.Error type that allows the caller to
further annotate the error more.

We have introduced a new Error type that allows annotating existing
errors with more information and tracing.

On top of this ConstError from juju/errors has been brought over with
the existing tests established.