Updates unique constraint on user name in table.
This commit updates the unique constraint on the user table for the
users name. The updated constraint allows for many users with the same
name to exist in the table as long as they are all deleted. The
constraint will maintain that only one non deleted user can exist in the
table with the same name at any time.

Tests have been included to assert this DDL logic and that a unique
constraint error is returned.