Insert factory for creation of objectstore
This builds on the refactoring from prior PRs that move implicit
dependencies to explicit dependencies for the object store. In order
to implement the file based object store we need to swap out the
object store from one to another at bootstrap. The factory for creating
the objectstore allows us to do that based on the default backend
type.

Additional work for controller-config holding the default backing
store type needs to be added.