Expose object store backend type in controller config
The following just exposes the backend object store type to the
controller conifg. This is currently only ever set at bootstrap,
so we don't need to wire this up to the agent config. Once s3 is
implemented we will require the objectstore type to be able to
change at runtime