Removal of lxd retry logic
The following was added in [1] to help with locating the vm if it
didn't start up in time. This isn't required any more as the vm
correctly starts up cleanly. In addition the retry was at level
that causes retries for requests that don't expect retries to
happen. This was identified in bug[2] as a failure.

One thing we should follow up on is the number of cores to run
a vm for bootstrap should be a minimum of 2.

 1. https://github.com/juju/juju/pull/15004
 2. https://bugs.launchpad.net/juju/+bug/2029524