Object store metadata schema
The object store will require tracking the underlying blobs of the
object store. To do this we need to keep track of the path, size
and hash. The path needs to be unique so multiple files are not
added, the hash is not unqiue as multiple metadata can point to
the same blob.