New password types and hasing for Juju and Dqlite.
This PR is mostly about consolidating the logic that we have in Juju
around user passwords into a new common package.

We are introducing a new password type for use in the code base. Because
Juju deals with plain text passwords over the wire from users we want a
safe way to wrap these raw strings up into a strong type. The reason for
doing this is to offer safety around working with the types. When the
password has been wrapped in the password type it should be almost
impossible for a developer to accidentally output the password in a log
or output stream.

We are also actively moving away from having lots of external Juju
libraries for offering utility like functions like password hashing.
Because of this we have moved the logic previously found in juju/utils
around password hashing into the core of Juju now.

Moving the code back into Juju offers us greater flexibility to changing
hashing algorithms and salting methods later on if we ever need to
upgrade.