Pass sample ratio flag
As discussed in person at the last sprint, we want to be able to
turn down the sample ratio. This is crude float64 value at the moment,
in the long run we should implement a type that can be decoded into
the sampling ruleset.

Luckily, we're starting to keep all the controller config values as
a string, so moving over to the new rule set would be easy.

The code just follows the existing open telemetry controller config
properties.