Add agent facades for the migration pre-check
The following forces that all agent facades are correctly registered
in the pre-source migration check. We currently don't block the check
yet, that will come in another PR. This just sets up registering them
and ensuring that there is a complete overlap between registered and
reported versions.

It's a shame that the registry doesn't give the information readily,
otherwise we could just use that!