Update deploy tests run on lxd.
One test was failing due to jammy not supported OS in the charm, update
the charm. Not caught as the test was in proving grounds and failed
before running due to nested containers not working.

Move nested container tests to providers other than lxd.

Specify operating system when deploying the
juju-qa-lxd-profile-without-devices charm. Xenial containers do not
always get past a pending state.

Update the basic openstack bundle to use focal-yoga as the current
charms do not support Xenial.