Fix LP 2003971.
The logic for handling a revision-not-found error from charmhub is wrong
and can result in deploying a charm to an unsupported operating system
when the operating system requested by the user isn't found in the
channel requested. Fix by always failing on that error rather than
checking to see if a user requested a specific channel for the charm.

This was missed as we never tested on a charm which had support for only
one operating system in a channel, where we tried to deploy a different
operating system. Adding a new test once we have a new channel for our
test charms.