Remove MAAS temp storage
This is a prototype to try and remove the legacy storage for
MAAS. According to the MAASServer API[1] we uploaded a file to
then retrieve at later date. This is some vestigial code from prior
to ssh'ing files to the instance. The MAAS team are looking to remove
this to clean up their API. As I don't believe we're using it anymore
we should be able to remove this from the maas provider.

The code just attempts to clean up any resources, but I don't believe
or can't see where we're actually putting anything into the storage. If
this is right, we should be able to just remove this code.

Testing will be required.

 1. https://git.launchpad.net/maas/tree/src/maasserver/api/files.py#n75