Wire up uniter with tracing
Add context.Context to all nextOp methods. That way we can start
passing the tracer to every layer. The uniter is a complex beast,
hence why we're doing it in multiple stages. The first being a
very mechanical code change.

The next set of changes will add tracting to the right levels. We
don't want a long trace, so we have to becareful where to add it.
If we start off small at first, we can start to understand some of
the requirements from introspecting the uniter.